* Shift from "Integer" to "Long" as per the new model * Updates to PixelBuffer to allow writes * Integration tests for pixel buffer creation and plane writing
git-svn-id: file:///home/svn/omero/trunk@303 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/omeio-nio/src/ome/io/nio/FileBuffer.java
@@ -67,7 +67,7 @@ public class FileBuffer
     // Delegate methods to ease work with original file
     //
     
-    int getId()
+    long getId()
     {
         return file.getId();
     }
